feat: Supply a HTML template for Preview
