+++ cypress/projects/themed/playroom.config.js
@@ -8,4 +8,5 @@ module.exports = {
   port: 9001,
   storageKey: 'playroom-example-themed',
   htmlTemplate: './template.ejs',
+  previewHtmlTemplate: './preview-template.ejs',
 };
