+++ cypress/projects/themed/preview-template.ejs
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8" />
+    <title><%= htmlWebpackPlugin.options.title %></title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="text/javascript">
+      console.info('Dummy code appended after Playroom Preview...');
+    </script>
+  </body>
+</html>
