+++ lib/makeWebpackConfig.js
@@ -163,6 +163,10 @@ module.exports = async (playroomConfig, options) => {
         chunksSortMode: 'none',
         chunks: ['preview'],
         filename: 'preview/index.html',
+        // Spread the template option, since setting it to null, undefined or "" if none is specified causes an issue.
+        ...(playroomConfig.previewHtmlTemplate
+          ? { template: relativeResolve(playroomConfig.previewHtmlTemplate) }
+          : {}),
       }),
       ...(options.production
         ? []
