+++ src/index.js
@@ -3,6 +3,7 @@ import { render } from 'react-dom';
 
 import Playroom from './Playroom/Playroom';
 import { StoreProvider } from './StoreContext/StoreContext';
+import getOrCreateRoot from './get-or-create-root';
 import playroomConfig from './config';
 
 const polyfillIntersectionObserver = () =>
@@ -10,34 +11,9 @@ const polyfillIntersectionObserver = () =>
     ? Promise.resolve()
     : import('intersection-observer');
 
-/**
- * Get the root where we want to render Playroom.
- * - If an `htmlTemplate` option is provided, use an element with ID "root" as the root.
- * - If no template is provided, simply create a `<div />` element and return it.
- */
-const getOrCreateRoot = () => {
-  if (playroomConfig.htmlTemplate) {
-    const root = document.getElementById('root');
-
-    if (!root) {
-      // If #root element is not found, throw error as we won't be able to render Playroom.
-      throw new Error(
-        'No element found in body with ID "root". Playroom won\'t be rendered. Put a `<div id="root"></div>` in your HTML template where you want to render Playroom.'
-      );
-    }
-
-    return root;
-  }
-
-  const outlet = document.createElement('div');
-  document.body.appendChild(outlet);
-
-  return outlet;
-};
-
 polyfillIntersectionObserver().then(() => {
   const widths = playroomConfig.widths || [320, 375, 768, 1024];
-  const root = getOrCreateRoot();
+  const root = getOrCreateRoot(playroomConfig.htmlTemplate);
 
   const renderPlayroom = ({
     themes = require('./themes'),
